Fix BUG while printing return value.

+++ syscall-replayer/SystemCallTraceReplayModule.hpp
@@ -196,7 +196,7 @@ public:
    * @return: corresponding return value field of a record that
    *          extent series is pointing to.
    */
-  int return_value() const;
+  int64_t return_value() const;
 
   /*
    * Get the unique id of current system call record
